Updates README to not use stack pointers on search
An issue with btree.h is that if a malloc'd struct is passed to it
it owns the pointer to the struct, but leaves the malloc'd members
on the heap.  When using `f_search_result_free` on a result stored
in the btree it will segfault when trying to free the allocated
search result. So there is a small memory leak when using btree
so I updated the README to pass a reference to the search result
pointer (instead of the pointer itself) when adding an item to the
btree.  I don't think this is the right call though, so reverting.